requirements.txt
allure_python_commons==2.12.0
allure_pytest==2.12.0
pytest==7.2.1
selenium==4.8.2
